requirements.txt
streamlit
python-docx
google-generativeai
python-dotenv
chromadb
sentence-transformers